@@ -1,6 +1,6 @@
 {
   "name": "simple-react-form",
-  "version": "1.4.1",
+  "version": "1.5.0",
   "license": "MIT",
   "author": "Nicolás López <nicolaslopezj@me.com>",
   "description": "Simply create forms using react",
@@ -33,7 +33,7 @@
     "uglify-js": "^2.6.1"
   },
   "peerDependencies": {
-    "react": ">=0.13.0"
+    "react": ">=15.0.0"
   },
   "scripts": {
     "build": "babel ./src --out-dir ./lib",
